make authentication directive dynamic authentication
