+++ packages/graphql/src/graphql/directives/type-dependant-directives/authentication.ts
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { astFromDirective } from "@graphql-tools/utils";
+import type { DirectiveDefinitionNode } from "graphql";
+import { GraphQLDirective, GraphQLInputObjectType, GraphQLList, DirectiveLocation } from "graphql";
+import { AUTHENTICATION_OPERATION } from "./static-definitions";
+
+const authenticationDefaultOperations = [
+    "READ",
+    "CREATE",
+    "UPDATE",
+    "DELETE",
+    "CREATE_RELATIONSHIP",
+    "DELETE_RELATIONSHIP",
+    "SUBSCRIBE",
+];
+function createAuthentication(jwtPayloadWhere: GraphQLInputObjectType): GraphQLDirective {
+    return new GraphQLDirective({
+        name: "authentication",
+        locations: [DirectiveLocation.OBJECT, DirectiveLocation.FIELD_DEFINITION],
+        args: {
+            operations: {
+                description: "operations",
+                type: new GraphQLList(AUTHENTICATION_OPERATION),
+                defaultValue: authenticationDefaultOperations,
+            },
+            jwtPayload: {
+                type: jwtPayloadWhere,
+            },
+        },
+    });
+}
+
+export function createAuthenticationDirectiveDefinition(): DirectiveDefinitionNode {
+    const jwtPayloadWhere = new GraphQLInputObjectType({ name: "JWTPayloadWhere", fields: {} });
+    const authentication = createAuthentication(jwtPayloadWhere);
+    const authenticationAST = astFromDirective(authentication);
+    return authenticationAST;
+}
