+++ packages/graphql/src/schema-model/annotation/AuthenticationAnnotation.ts
@@ -17,6 +17,8 @@
  * limitations under the License.
  */
 
+import type { GraphQLWhereArg } from "../../types";
+
 export type AuthenticationOperation =
     | "READ"
     | "CREATE"
@@ -28,12 +30,18 @@ export type AuthenticationOperation =
 
 export class AuthenticationAnnotation {
     private _operations: AuthenticationOperation[];
+    private _jwtPayload?: GraphQLWhereArg;
 
-    constructor(operations: AuthenticationOperation[]) {
+    constructor(operations: AuthenticationOperation[], jwtPayload?: GraphQLWhereArg) {
         this._operations = operations;
+        this._jwtPayload = jwtPayload;
     }
 
     public get operations() {
         return this._operations;
+    }
+
+    public get jwtPayload() {
+        return this._jwtPayload;
     }
 }
