+++ packages/graphql/src/schema-model/parser/authentication-annotation.ts
@@ -17,6 +17,7 @@
  * limitations under the License.
  */
 import type { DirectiveNode } from "graphql";
+import type { GraphQLWhereArg } from "../../types";
 import type { AuthenticationOperation } from "../annotation/AuthenticationAnnotation";
 import { AuthenticationAnnotation } from "../annotation/AuthenticationAnnotation";
 
@@ -33,11 +34,15 @@ const authenticationDefaultOperations: AuthenticationOperation[] = [
 export function parseAuthenticationAnnotation(directive: DirectiveNode): AuthenticationAnnotation {
     const args = parseArguments(directive) as {
         operations?: AuthenticationOperation[];
+        jwtPayload?: GraphQLWhereArg;
     };
 
-    if (args.operations) {
-        return new AuthenticationAnnotation(args.operations);
+    const constructorArgs: [AuthenticationOperation[], GraphQLWhereArg?] = [
+        args.operations || authenticationDefaultOperations,
+    ];
+    if (args.jwtPayload) {
+        constructorArgs.push(args.jwtPayload);
     }
 
-    return new AuthenticationAnnotation(authenticationDefaultOperations);
+    return new AuthenticationAnnotation(...constructorArgs);
 }
